third_party/blink/web_tests/platform/generic/http/tests/websocket/workers/send-arraybufferview-expected.txt

WebSocket: Send ArrayBufferViews in Web Workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PASS: Message #0.
PASS PASS: Message #1.
PASS PASS: Message #2.
PASS PASS: closeEvent.wasClean is true.
DONE
PASS successfullyParsed is true

TEST COMPLETE
